+++ config/sql/rmgdb/schema_23_up.sql
@@ -1,5 +1,5 @@
 BEGIN;
 
-ALTER TABLE transaction.transaction ADD COLUMN markup INT DEFAULT 0;
+ALTER TABLE transaction.transaction ADD COLUMN markup NUMERIC(1,2) DEFAULT 0.00;
 
 COMMIT;
